refactor(api): extract rate limiter into internal/api/ratelimit leaf package
Second internal/api decomposition step (ADR-0006). The per-IP RateLimiter
(map + FIFO eviction + stale sweep) and the four per-endpoint middleware
wrappers move into internal/api/ratelimit.

The wrappers were *Server methods reaching into s.{upload,write,walk,file}Limiter,
s.logger, and the unexported getClientIP / writeError. They become free factory
functions — ratelimit.Upload/Write/Walk/File — taking the limiter, logger, and
two injected function values: ClientIPFunc and a narrow ErrorFunc. ErrorFunc
deliberately omits writeError's []ErrorDetail param (rate-limit rejections carry
no details), so the leaf imports no api type; route.go passes a small
rateLimitErr adapter. MaxRateLimiterCount moves with the limiter.

- route.go's register() switch now composes ratelimit.<Kind>(s.<x>Limiter,
  s.logger, getClientIP, rateLimitErr, h); the canonical middleware order and
  /__capabilities are unchanged.
- Tests that touch unexported fields (rate/burst/limiters/mu, entry.lastSeen for
  stale/eviction simulation) move to package ratelimit; getClientIP-dependent and
  Server-level tests stay in package api via the exported surface.
- depguard gains api-ratelimit-isolated (the leaf may not import internal/api).

No behavioural change. Verified: go build, go vet, golangci-lint (0 issues),
go test -race ./internal/api/... ./internal/daemon/... (all pass), gofumpt clean,
route-policy gate green.
