+++ .golangci.yml
@@ -256,6 +256,12 @@ linters:
           deny:
             - pkg: github.com/MustardSeedNetworks/niac-go/internal/api
               desc: "tokenstore is a leaf — it holds the bearer-token store and scope model with no transport dependency; the api package imports it, never the reverse"
+        "api-ratelimit-isolated":
+          files:
+            - "**/internal/api/ratelimit/**"
+          deny:
+            - pkg: github.com/MustardSeedNetworks/niac-go/internal/api
+              desc: "ratelimit is a leaf — per-IP limiter + middleware factories take injected ClientIPFunc/ErrorFunc, so it never imports the api transport layer"
 
     embeddedstructfieldcheck:
       # Checks that sync.Mutex and sync.RWMutex are not used as embedded fields.
